requirements.txt
cmd2>=2.0.0
prompt_toolkit>=3.0.0
paramiko>=2.7.0
smolagents>=0.1.0
openai>=1.0.0
pyyaml>=6.0
